requirements.txt
django==3.2
Pillow
mysqlclient 
django-admin-tailwind
django-paypal
python-decouple
